reorganized admin code
