+++ src/app/admin/ownership/ownership.component.html
