+++ src/app/app-routing.module.ts
@@ -25,18 +25,18 @@ import { CreateNewProjectComponent } from './organizations/organization-details/
 import { ManageProjectsComponent } from './organizations/organization-details/manage-projects/manage-projects.component';
 import { ManageSingleProjectComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-single-project.component';
 import { VerifySignOfferComponent } from './offers/offer-details/verify-sign-offer/verify-sign-offer.component';
-import { ManageWithdrawalsComponent } from './manage-withdrawals/manage-withdrawals.component';
-import { SingleWithdrawalComponent } from './manage-withdrawals/single-withdrawal/single-withdrawal.component';
-import { ManageDepositsComponent } from './manage-deposits/manage-deposits.component';
+import { ManageWithdrawalsComponent } from './admin/manage-withdrawals/manage-withdrawals.component';
+import { SingleWithdrawalComponent } from './admin/manage-withdrawals/single-withdrawal/single-withdrawal.component';
+import { ManageDepositsComponent } from './admin/manage-deposits/manage-deposits.component';
 import { DepositComponent } from './deposit/deposit.component';
-import { ManageSingleDepositComponent } from './manage-deposits/manage-single-deposit/manage-single-deposit.component';
+import { ManageSingleDepositComponent } from './admin/manage-deposits/manage-single-deposit/manage-single-deposit.component';
 import { WithdrawComponent } from './withdraw/withdraw.component';
-import { WalletActivationComponent } from './wallet-activation/wallet-activation.component';
+import { WalletActivationComponent } from './admin/wallet-activation/wallet-activation.component';
 import { CompleteOnboardingComponent } from './complete-onboarding/complete-onboarding.component';
-import { PlatformBankAccountComponent } from './platform-bank-account/platform-bank-account.component';
-import { NewPlatformBankAccountComponent } from './platform-bank-account/new-platform-bank-account/new-platform-bank-account.component';
+import { PlatformBankAccountComponent } from './admin/platform-bank-account/platform-bank-account.component';
+import { NewPlatformBankAccountComponent } from './admin/platform-bank-account/new-platform-bank-account/new-platform-bank-account.component';
 import { ExchangeComponent } from './exchange/exchange.component';
-import { OwnershipComponent } from './ownership/ownership.component';
+import { OwnershipComponent } from './admin/ownership/ownership.component';
 import { RevenueShareComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-payments/revenue-share/revenue-share.component';
 import { ManagePaymentsComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-payments/manage-payments.component';
 import { OfferDetailsGuard } from './offers/offer-details/offer-details.guard';
