+++ src/app/app.module.ts
@@ -9,14 +9,12 @@ import { AppRoutingModule } from './app-routing.module';
 import { WalletComponent } from './wallet/wallet.component';
 import { OffersComponent } from './offers/offers.component';
 import { SingleOfferItemComponent } from './offers/single-offer-item/single-offer-item.component';
-import { WalletTxHistoryComponent } from './wallet/wallet-tx-history/wallet-tx-history.component';
 import { OfferDetailsComponent } from './offers/offer-details/offer-details.component';
 import { DepositModalComponent } from './wallet/deposit-modal/deposit-modal.component';
 import { InvestComponent } from './invest/invest.component';
 import { MyPortfolioComponent } from './my-portfolio/my-portfolio.component';
 import { SingleInvestItemComponent } from './my-portfolio/single-invest-item/single-invest-item.component';
 import { InvestmentDetailsComponent } from './investment-details/investment-details.component';
-import { SingleProposalItemComponent } from './investment-details/single-proposal-item/single-proposal-item.component';
 import { PaymentOptionsComponent } from './payment-options/payment-options.component';
 import { FooterComponent } from './footer/footer.component';
 import { NewProposalComponent } from './investment-details/new-proposal/new-proposal.component';
@@ -44,22 +42,22 @@ import { ManageProjectsComponent } from './organizations/organization-details/ma
 import { ManageSingleProjectComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-single-project.component';
 import { VerifySignOfferComponent } from './offers/offer-details/verify-sign-offer/verify-sign-offer.component';
 import { ManagePaymentsComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-payments/manage-payments.component';
-import { ManageWithdrawalsComponent } from './manage-withdrawals/manage-withdrawals.component';
-import { SingleWithdrawalComponent } from './manage-withdrawals/single-withdrawal/single-withdrawal.component';
+import { ManageWithdrawalsComponent } from './admin/manage-withdrawals/manage-withdrawals.component';
+import { SingleWithdrawalComponent } from './admin/manage-withdrawals/single-withdrawal/single-withdrawal.component';
 import { DepositComponent } from './deposit/deposit.component';
-import { ManageDepositsComponent } from './manage-deposits/manage-deposits.component';
-import { ManageSingleDepositComponent } from './manage-deposits/manage-single-deposit/manage-single-deposit.component';
+import { ManageDepositsComponent } from './admin/manage-deposits/manage-deposits.component';
+import { ManageSingleDepositComponent } from './admin/manage-deposits/manage-single-deposit/manage-single-deposit.component';
 import { WithdrawComponent } from './withdraw/withdraw.component';
-import { WalletActivationComponent } from './wallet-activation/wallet-activation.component';
-import { UserActivationComponent } from './wallet-activation/user-activation/user-activation.component';
-import { GroupActivationComponent } from './wallet-activation/group-activation/group-activation.component';
-import { ProjectActivationComponent } from './wallet-activation/project-activation/project-activation.component';
+import { WalletActivationComponent } from './admin/wallet-activation/wallet-activation.component';
+import { UserActivationComponent } from './admin/wallet-activation/user-activation/user-activation.component';
+import { GroupActivationComponent } from './admin/wallet-activation/group-activation/group-activation.component';
+import { ProjectActivationComponent } from './admin/wallet-activation/project-activation/project-activation.component';
 import { CompleteOnboardingComponent } from './complete-onboarding/complete-onboarding.component';
 import { ModalModule } from 'ngx-bootstrap/modal';
-import { PlatformBankAccountComponent } from './platform-bank-account/platform-bank-account.component';
-import { NewPlatformBankAccountComponent } from './platform-bank-account/new-platform-bank-account/new-platform-bank-account.component';
+import { PlatformBankAccountComponent } from './admin/platform-bank-account/platform-bank-account.component';
+import { NewPlatformBankAccountComponent } from './admin/platform-bank-account/new-platform-bank-account/new-platform-bank-account.component';
 import { ExchangeComponent } from './exchange/exchange.component';
-import { OwnershipComponent } from './ownership/ownership.component';
+import { OwnershipComponent } from './admin/ownership/ownership.component';
 import { CurrencyDefaultPipe } from './shared/pipes/currency-default.pipe';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TxAmountSign, TxIconStatus, TxIconType } from './wallet/wallet.pipe';
@@ -81,7 +79,7 @@ import { SafePipe } from './shared/pipes/safe.pipe';
 import { URLValidator } from './shared/validators/url.validator';
 import { ForgotPasswordComponent } from './authentication/forgot-password/forgot-password.component';
 import { ResetPasswordComponent } from './authentication/reset-password/reset-password.component';
-import { ManageSingleDepositModalComponent } from './manage-deposits/manage-single-deposit/manage-single-deposit-modal/manage-single-deposit-modal.component';
+import { ManageSingleDepositModalComponent } from './admin/manage-deposits/manage-single-deposit/manage-single-deposit-modal/manage-single-deposit-modal.component';
 import { RevenueShareComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-payments/revenue-share/revenue-share.component';
 import { RevenueShareConfirmModalComponent } from './organizations/organization-details/manage-projects/manage-single-project/manage-payments/revenue-share/revenue-share-confirm-modal/revenue-share-confirm-modal.component';
 import { ProjectDepositComponent } from './organizations/organization-details/manage-projects/manage-single-project/project-deposit/project-deposit.component';
@@ -127,14 +125,12 @@ const socialAuthServiceConfig = {
         WalletComponent,
         OffersComponent,
         SingleOfferItemComponent,
-        WalletTxHistoryComponent,
         OfferDetailsComponent,
         DepositModalComponent,
         InvestComponent,
         MyPortfolioComponent,
         SingleInvestItemComponent,
         InvestmentDetailsComponent,
-        SingleProposalItemComponent,
         PaymentOptionsComponent,
         FooterComponent,
         NewProposalComponent,
