--- src/app/investment-details/single-proposal-item/single-proposal-item.component.html
@@ -1,17 +1,0 @@
-<div class="card card-app">
-    <div class="card-body px-5 py-5">
-        <h5 class="text-uppercase ">{{ proposal.name }}</h5>
-        <p class="text-justify">
-            {{ proposal.description }}
-        </p>
-        <div class="dropdown-divider mt-3 mb-2"></div>
-
-        <div class="progress mt-3">
-            <div aria-valuemax="100" aria-valuemin="0" aria-valuenow="25" class="progress-bar" role="progressbar"
-                 style="width: 25%"></div>
-        </div>
-        <p class="mt-3 explainer-mini text-uppercase"><i>27%</i> votes positive; <i>377</i> votes (<i>20%</i> weighted)
-            so far</p>
-        <button class="btn btn-primary float-right mt-2" routerLink="proposal_details">Details</button>
-    </div>
-</div>
