--- src/app/models/OfferDetailDocModel.ts
@@ -1,9 +1,0 @@
-export interface OfferDetailDocModel {
-    docType: OfferDetailDocType;
-    title: string;
-    src: URL;
-}
-
-export enum OfferDetailDocType {
-    PDF, DOC, XLS, PNG, JPG, OTHER
-}
