--- src/app/models/PaymentOptionModel.ts
@@ -1,9 +1,0 @@
-export interface PaymentOptionModel {
-    name: string;
-    type: PaymentOptionType;
-    active: boolean;
-}
-
-export enum PaymentOptionType {
-    bankAccount, creditCard
-}
