--- src/app/models/ProposalModel.ts
@@ -1,7 +1,0 @@
-export interface ProposalModel {
-    name: string;
-    description: string;
-    positiveVotes: number;
-    negativeVotes: number;
-    allVotesCount: number;
-}
