--- src/app/models/TransactionHistoryModel.ts
@@ -1,13 +1,0 @@
-export interface TxHistoryModel {
-    position: number;
-    date: Date;
-    amount: number;
-    type: TxHistoryType;
-}
-
-export enum TxHistoryType {
-    Deposit,
-    Withdraw,
-    Send,
-    Receive
-}
