--- src/app/models/auth/TokenModel.ts
@@ -1,6 +1,0 @@
-export class TokenModel {
-    access_token: string;
-    expires_in: number;
-    refresh_token: string;
-    refresh_token_expires_in: number;
-}
