--- src/app/models/auth/check-email-model.ts
@@ -1,4 +1,0 @@
-export class CheckEmailModel {
-    email: string;
-    userExists: boolean;
-}
