--- src/app/models/payment-model.ts
@@ -1,11 +1,0 @@
-export interface PaymentModels {
-    bank_accounts: SinglePaymentModel[];
-}
-
-export interface SinglePaymentModel {
-    id: number;
-    iban: string;
-    bank_code: string;
-    created_at: string;
-    alias: string;
-}
