--- src/app/models/single-organization-model.ts
@@ -1,9 +1,0 @@
-export class SingleOrganizationModel {
-    uuid: number;
-    name: string;
-    created_by_user: string;
-    created_at: string;
-    approved: boolean;
-    legal_info: string;
-    documents: string[];
-}
