--- src/app/models/user-model.ts
@@ -1,8 +1,0 @@
-// export interface UserModel {
-//     uuid: string;
-//     email: string;
-//     first_name: string;
-//     last_name: string;
-//     enabled: boolean;
-//     verified: boolean;
-// }
