--- src/app/wallet-activation/group-activation/group-activation.model.ts
@@ -1,10 +1,0 @@
-import { OrganizationModel } from 'src/app/organizations/organization-details/organization-model';
-
-export interface GroupActivationModel {
-    organization: OrganizationModel;
-    wallet: GroupActivationWalletModel;
-}
-
-interface GroupActivationWalletModel {
-    uuid: string;
-}
