--- src/app/wallet-activation/project-activation/project-activation.model.ts
@@ -1,8 +1,0 @@
-export interface ProjectActivationModel {
-    organization: any;
-    wallet: ProjectActivationWalletModel;
-}
-
-interface ProjectActivationWalletModel {
-    uuid: string;
-}
