--- src/app/wallet-activation/user-activation/user-activation.model.ts
@@ -1,16 +1,0 @@
-export interface UserActivationModel {
-    user: UserActivationUserSubmodel;
-    wallet: UserActivationWalletModel;
-}
-
-interface UserActivationUserSubmodel {
-    uuid: string;
-    email: string;
-    first_name: string;
-    last_name: string;
-    enabled: boolean;
-}
-
-interface UserActivationWalletModel {
-    uuid: number;
-}
