--- src/app/wallet/wallet-tx-history/wallet-tx-history.component.css
@@ -1,8 +1,0 @@
-.history-row {
-    font-size: 14px;
-    color: #555;
-}
-
-.history-row:hover {
-    cursor: pointer;
-}
