--- src/app/wallet/wallet-tx-history/wallet-tx-history.component.html
@@ -1,91 +1,0 @@
-<!-- <div class="">
-	<div class="list-group-item">
-		<div class="row p-2 text-center explainer-mini text-uppercase">
-			<div class="col-md-3">
-				<i class="fas fa-exchange-alt mr-2"></i>
-				Type
-			</div>
-			<div class="col-md-3">
-				<i class="fas fa-calendar mr-2"></i>
-				Date
-			</div>
-			<div class="col-md-3">
-				<i class="fas fa-money-bill-alt mr-2"></i>
-				Amount
-			</div>
-			<div class="col-md-3">
-				<i class="fas fa-id-card mr-2"></i>
-				TX ID
-			</div>
-		</div>
-	</div>
-	<button *ngFor="let historyItem of txHistoryModel"
-		class="list-group-item list-group-item-action history-row"
-		routerLink="tx_overview">
-		<div class="row px-2 py-1 text-center  ">
-			<div class="col-md-3">
-				<div *ngIf="historyItem.type ===  0">
-					<i class="fas fa-arrow-circle-down mr-3"></i>
-					<span>DEPOSIT</span>
-				</div>
-				<div *ngIf="historyItem.type ===  1">
-					<i class="fas fa-arrow-circle-down mr-3"></i>
-					<span>DEPOSIT</span>
-				</div>
-			</div>
-			<div class="col-md-3">
-				<b>SEP</b> 19 2018
-			</div>
-			<div class="col-md-3">
-				USD {{ historyItem.amount }}
-			</div>
-			<div class="col-md-3">
-				<a href="http://etherscan.io">0xABB3C72F53489F3937FG</a>
-			</div>
-		</div>
-	</button>
-</div> -->
-<div>
-    <div class="row p-2 text-center explainer-mini text-uppercase mb-3">
-        <div class="col-md-3">
-            <i class="fas fa-exchange-alt mr-2"></i>
-            Type
-        </div>
-        <div class="col-md-3">
-            <i class="fas fa-calendar mr-2"></i>
-            Date
-        </div>
-        <div class="col-md-3">
-            <i class="fas fa-money-bill-alt mr-2"></i>
-            Amount
-        </div>
-        <div class="col-md-3">
-            <i class="fas fa-id-card mr-2"></i>
-            TX ID
-        </div>
-    </div>
-    <div *ngFor="let tx of txHistoryModel" class="row text-center" style="font-size: 12px">
-        <div class="col-md-3">
-            <div *ngIf="tx.type ===  0">
-                <i class="fas fa-arrow-circle-down mr-3"></i>
-                <span>DEPOSIT</span>
-            </div>
-            <div *ngIf="tx.type ===  1">
-                <i class="fas fa-arrow-circle-down mr-3"></i>
-                <span>DEPOSIT</span>
-            </div>
-        </div>
-        <div class="col-md-3">
-            <b>SEP</b> 19 2018
-        </div>
-        <div class="col-md-3">
-            USD {{ tx.amount }}
-        </div>
-        <div class="col-md-3">
-            <a href="http://etherscan.io">0xABB3C72F53489F3937FG</a>
-        </div>
-        <div class="col-md-12 px-5 mt-2">
-            <div class="dropdown-divider"></div>
-        </div>
-    </div>
-</div>
